feat(LottoController): 로또를 실행할 수 있다.
