+++ src/main/java/lotto/controller/LottoController.java
@@ -1,0 +1,18 @@
+package lotto.controller;
+
+import lotto.view.InputView;
+import lotto.view.OutputView;
+
+public class LottoController {
+    private final InputView inputView;
+    private final OutputView outputView;
+
+    public LottoController(InputView inputView, OutputView outputView) {
+        this.inputView = inputView;
+        this.outputView = outputView;
+    }
+
+    public void run() {
+
+    }
+}
